🐛 Fix the issue where the Chinese language is displayed after refreshing the English interface.
